:recycle: Extract Maven settings code to maven_settings.go
Move all Maven globalSettings.xml types, parsing, building, marshaling,
BuildSettingsFile, and redactProxyURL from provider.go into a dedicated
maven_settings.go file. Tests move to maven_settings_test.go.

Also use 0755/0644 permissions for the directory and file, and check
MkdirAll errors in tests.
